Move messages into own class and interface
